service_SQL/requirements.txt
Flask_SQLAlchemy==2.4.0
Flask==1.0.2
PyYAML==5.1.1
